Added the functionality of getting the week schedule of the bookings into the LaundryDAO and all the needed components for the functionality
Added two functions getResidentPasses() and getPasses() To allow the standard and administrators to get the schedule in a separate manner. Added two DTOs PassDTO and PassScheduleDTO to hold the information about the bookings. In addition to some enumerators for the schedule status codes and slot status. Added missing check in the cancelBookedPass().